+++ Source/Fonts/Extensions/UIFont+ScaledFont.swift
@@ -34,7 +34,7 @@ extension UIFont {
 
      - Returns: A scalable font object.
      */
-    public class func scaledFont(name fontName: String, size fontSize: CGFloat) -> UIFont? {
+    public class func scaledFont(withName fontName: String, ofSize fontSize: CGFloat) -> UIFont? {
         if let font = UIFont(name: fontName, size: fontSize) {
             let fontMetrics = FontMetrics()
             return fontMetrics.scaledFont(for: font)
@@ -46,7 +46,7 @@ extension UIFont {
      Creates a scalable system font with a given reference font size.
 
      - Parameters:
-     - fontSize: The reference font size to use.
+        - fontSize: The reference font size to use.
 
      - Returns: A scalable font object.
      */
@@ -59,7 +59,7 @@ extension UIFont {
      Creates a scalable bold system font with a given reference font size.
 
      - Parameters:
-     - fontSize: The reference font size to use.
+        - fontSize: The reference font size to use.
 
      - Returns: A scalable font object.
      */
@@ -72,7 +72,7 @@ extension UIFont {
      Creates a scalable italic system font with a given reference font size.
 
      - Parameters:
-     - fontSize: The reference font size to use.
+        - fontSize: The reference font size to use.
 
      - Returns: A scalable font object.
      */