+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## Version 0.9.0
+
+### Features
+* [Add WCAG colors extension](https://github.com/chrs1885/Capable/issues/39)
+* [Add feature hearingDevice on iOS](https://github.com/chrs1885/Capable/issues/38)
+
+### Enhancements
+* [CocoaPods subspec for UIFontMetrics/UIFont extension](https://github.com/chrs1885/Capable/issues/35)
+
 ## Version 0.8.0
 
 ### Features